+++ scripts/maxcut_random_sparse.py
@@ -13,7 +13,7 @@ def generate_adjacency(n_nodes=64, seed=None):
     if not (seed is None):
         np.random.seed(seed)
 
-    lil = lil_matrix((n_nodes, n_nodes), dtype=np.float64)
+    lil = lil_matrix((n_nodes, n_nodes), dtype=np.float32)
 
     for u in range(n_nodes >> 1):
         for v in range(u + 1, n_nodes >> 1):
